restructured project, created own app for drinks
