+++ tally/tally/settings.py
@@ -37,8 +37,9 @@ INSTALLED_APPS = [
     "django.contrib.sessions",
     "django.contrib.messages",
     "django.contrib.staticfiles",
-    "api",
     "rest_framework",
+    "api",
+    "drinks",
 ]
 
 MIDDLEWARE = [
